Split virtual environment detection into a dedicated module
# Conflicts:
#	crates/uv-interpreter/src/find_python.rs
