--- crates/uv-interpreter/src/environment/mod.rs
@@ -1,3 +1,0 @@
-pub(crate) mod cfg;
-pub(crate) mod python_environment;
-pub(crate) mod virtualenv;
