--- crates/uv-interpreter/src/environment/virtualenv.rs
@@ -1,17 +1,0 @@
-use std::path::PathBuf;
-
-use pypi_types::Scheme;
-
-/// The layout of a virtual environment.
-#[derive(Debug)]
-pub struct Virtualenv {
-    /// The absolute path to the root of the virtualenv, e.g., `/path/to/.venv`.
-    pub root: PathBuf,
-
-    /// The path to the Python interpreter inside the virtualenv, e.g., `.venv/bin/python`
-    /// (Unix, Python 3.11).
-    pub executable: PathBuf,
-
-    /// The [`Scheme`] paths for the virtualenv, as returned by (e.g.) `sysconfig.get_paths()`.
-    pub scheme: Scheme,
-}
